Refactor UserProfileCalendar component
The impetus of this was to fix the component's data loading on Sunday
night in en-us where the data from the previous week would get loaded
instead. I replaced the loading modifier with a tracked async getter and
linked this up to the localeDays service.

I also replaced the loader with the built in one that is now in the
weekly calendar.