rm scroll-into-view modifier with did-insert based equivalent.
the initial idea was to remove yet another building block that uses
runloop functionality directly, since we're actively linting against
that now.
on further inspection, it turns out that the modifier based
implementation "jumps" rather than smoothly scrolls to the element in
question. we're using the smooth-scrolling solution elsewhere already,
so let's consolidate around that.

this required some markup to move in one instance. other than that, this is
a drop-in replacement.