Fix staging deploy runs
Using the GITHUB_TOKEN to create an automerge will not trigger other
github actions so our staging deploys aren't running when these get
merged. Github recommends we use a PAT, and we've got one ready made
from Zorgbort. This was Zorgbort is responsible for the merge and github
is happy to go forth and trigger more actions.