Declare pnpm in package
We need to do this in order to work around an issue with dependabot that
is using an old pnpm version. The downside is we'll need to keep it up
to date. My IDE is telling me this format isn't valid and it needs to be
x.y.z format, but let's try it with just the major version and see if
that solves our issues as we won't need to keep it up to date this way.