Replace element modifier
Refactoring the data loading here so we don't have to do it when the
element is inserted into the DOM. Because this data is now loaded and
re-loaded on each change I needed to pay some tricks to show the right
values during that process.

I also re-wrote the save method here while I was trying to solve display
issues, I like it, gonna keep it.