Move percy back into CI
We've been playing some tricks to lower our screenshot count and not
expose the percy token, however it's not working as we rarely remember
to add this label. I'm tossing all the previous ideas and going back to
running percy in CI.

I did add a check so that it won't run on draft pull requests, but will
run everywhere else.