Replace Reactivity with Constructor Assignment
When the arguments to this component change it's going to get torn down
so we don't need to worry about being reactive. However, the solution
I'd created here doesn't play well with our validators so going back to
this is better. I added a test to catch this next time.