Get Review State from the Github API
Using the mergable key didn't work, even though we required one review
for a PR it was still deemed mergable. Instead I'm using the github API
to determine if there is an approving review. It's not entirely clear
what will happen if there is also a change requested review, but I think
this is good enough for our purposes here.