Scroll Calendar Without Scrolling the Page
Instead of scrolling the calendar element into view I've used the
scrolling container element itself to scroll to the position of the
first event. This happens anytime the events list is updated.

The offsetTop property on an element uses a feature called offsetParent
which discovers the nearest ancestor element with a relative or absolute
position. I've added that positioning property to the scrolling
container so it is correctly identified as the offsetParent.