Replace Approval Based UI Workflow
This process isn't working for us. Because many people can approve a PR
it often runs many times and the final approval by QA often isn't
mergable. Instead I'm tying running the UI jobs on to a label.

This label can be applied at any time and is required for a PR to be merged.