Optimize the Instructors for Course Report
Because of the joins required this report performed very poorly using
the API as the number of sessions and offerings in a course grew.
Instead we can use our GraphQL API to get the data we want from the
course down and significantly improve the performance.

I didn't DRY this up, we probably need to replace the reporting service
for download and re-use the data fetching in the subject components, but
that is for another day.