Update Some Automation Timings
After a few weeks of running I see some adjustments we can make here to
smooth out this process. We need to run the percy nightly job on Sunday
so our Transitive Dependency job gets the latest screenshots, we shoudl
run dependabot on Monday so we don't waste time when the auto update
would have caught all the updates, and we can push back running the
transitive dependency update job a few hours.