More Improvements to Calendar Courses Filter
By marking the filter parent as relative positioned we can lookup the
distance of each year from the top of that filter and scroll the correct
one into view without opening anything else.

I discovered while testing this that if the current academic year
doesn't have any course we were opening the oldest year instead of the
latest year with courses. That seems wrong to me so I modified that
behavior as well.