Configure lcov Reporter projectRoot
Passing this will tell istanbul to create the lcov report from the
perspective of the monorepo root instead of from inside the frontend
directory (where we run the tests). This will allow us to hook up the
coverage results with the full path to the code (which is how
codeclimate sees our source).