Refactor yup validation translations
The initializer was a no go, it blew up tests. Plus this was everything
is in the same file. When the file is loaded so are the error message
translations.

We extract any needed information for the translation from the error and
send the entire thing to the message component. That component then does
the translation and subs in the description when needed.