Ignore Scripts in Auto Update Workflow
The scripts is where husky installs the precommit hook and the
recommended approach to skipping verification of this commit is to skip
installing that hook. The resulting PR will fail validation all on it's
own, but at least the error is fixable there. When it fails on commit in
this job that failure is hidden.