Replace Data Loading in Material Manager
Because this component is only ever rendered with a material and a
different material cannot be passed in we can do all of the data loading
in the constructor and not worry about the behavior in did-update.

I was also able to remove the spinner as the only data included in the
manager that isn't already loaded is some mesh tree meta data when mesh
terms are attached.