use an empty object to track course de-selection by the user.
all this just to get the course details of the current course selection - if there is one - , or the course already associated with the sequence block, correctly onto the screen.
i hate this, but can't come up with a better way of handling this multi-state situation.
speaking of which - the course details should probably be its own component now, will deal with that later.