decouples start/end academic level init and state tracking from load method()
form validation is hinky for these two, as it was before.
there's a slight change in behaviour here - if both are in edit-mode and
you save one, the other one closes as well. infuriating!
might as well lock one down while the other is in edit-mode.
that's probably going to be the next commit.