Run a courses report
Pull all the data from the GraphQL API for all of these selected courses
and put a summary on the screen.

There are some guesses here. I've selected the first offering of ILM to
determine start date, for example. Will have to test the output and see
if it's correct.