rm loading state check that was causing start/end level task chaining problems.
turns out this is both unnecessary and causing new trouble.
the issue here is that the start/end academic level getter chain runs as soon as a new value is set on the given sequence block, causing the in-place editor component from being unloaded instantly, thus aborting the save task chain. madness!