replace render modifiers and load() with chained getters equivalent.
this refactoring introduces an new, minor kink with form validation -
starting and ending academic year validation does not seem to kick in
until the user hits the save button. after that, it becomes responsive.
couldn't figure out where the problem exactly is with this, but
considering that a) the CI report subsystem is essentially headed for
sunsetting, and b) we're moving on to a less hinky form validation
library, i think that this is (barely) acceptable.
on the upside - i added validation error output to the starting academic
year element, and added a testcase for start/end academic year
validation to the component's integration test.