Fix weeksInWeekYear calculation
Doing this by modifying the current year didn't work. I'm not 100% sure
why, but the right way to run a single value into a Luxon DateTime is
with fromObject so I switched to this.

This prevents us from passing a bad combination of weeks in a year and
the year and blowing up the page.