Configure Our Session Cookie Better
This `ilios-session` cookie is used to store internal frontend
authentication data. We don't use it to authenticate against the
backend, we pass the `X-JWT-Authorization` header for that. As it's only
ever used on the frontend, where it is generated, we should be able to
restrict it to that domain (Strict). This has the added benefit of
reducing our payload on each request as we're no longer sending our JWT
twice.