better track given locale and locale-changes.
the flatpickr locale handling is inconsistent - if you give it a locale
identifier, like we do by default with `en`, and then try to get that
back from the picker instance, you'll get an object.
handling it this way is much cleaner, at the expense of adding another property.
it also gets us around having to compare object to object for
non-english locales.