Handle More Error Types
Depending on where it is thrown and how it transits the app sometimes
ember-data errors don't have the data we'd expect. This seems to be a
problem when a loading route is invoked specifically. It's too bad we
can't rely on the status code for these, but not found pretty clear
points to a 404, so let's key on that.