Fix Courses by Mesh Term Report
The SQL query we need to run on the API has way too many joins and
performs terribly for the courses by meshDescriptors filter. Instead we
can reverse the lookup here and first pull the descriptor and all of
it's relationships and then turn that into course ids and query them
like we do any other filter.

Makes for messy tests because the graphQL responses have to use a
counter, but it's 100x faster and won't crash the API DB anymore so that
seems worth it.