Downgrade @ember/string
The v4.0.x release of @ember/string is no longer an ember addon. This is
mostly fine, however when it is consumed from another addon (like
ember-responsive) this causes build issues. While we await updates to
those addons we can go back a version here.

I've unlocked this in dependabot so we'll know when it is fixed.