Use Component to Replace Ember Responsive Colspans
There isn't a reasonable way to set colspan to cover the entire table as
rendered so I'm doing this weird thing. I create two td elements with
the correct dimensions for different table sizes and then I hide the one
we don't need. Is it weird.? Yes. Does it work? Also yes.