Limit Visible Test in Publication Status
Instead of displaying the status as icon text I've introduced a new
option to FaIcon which allows us to specify the label for an icon
instead of having it be part of the icon itself.

Now the icon can display and be part of the dom tree, but not have a
visible title that gets picked up in tests.