Use a Modifier for Validation
We had several different patterns here to both start showing the error
message and to ensure that the message gets updated when the value
changes. I've replaced them all with a modifier on the validation class
that doesn't show any errors until the user focuses out of the input,
but does re-run the validations anytime the value changes.