Use Corepack to Determine pnpm Version
Using the packageManager field in our package.json file activates
corepack from Node and ensures we're always in alignment on what version
of pnpm we're using and what is used in our workflows.

I've added a workflow that will update this to the latest version once a
week.

I initially set the version to 10.0.0 so we'll have something to update
right away.