Improve Handling of Values in Yup
We use the `values` key to pass information to the translation, it
should always be an object and if it's not set (for instance in a `test`
type of validation where it's developer specified) we should account for
that and not expect it to be present. This is what we do in the
validations message component.