Fix Whitespace in Icon
We need to un-prettier this component so whitespace doesn't creep in
around the icon. When icons are used in a link the extra space can
create a very ugly underline in some places. We have a test for this
from the original ember-awesome home of this component and the behavior
seems worth keeping.