use Yup to validate program-year/objective-list-item component.
couple of things going on here.

1. the @HtmlNotBlank decorator has been replaced with the stripHtml()
   function. Yup validates against a wrapper getter around the actual
title attribute. good enough.
2. I yanked the code that was responsible for disabling the submit
   button in the edit-in-place component since we lost reactiveness
here. I tried to get there via an "hasErrorFor()" in the YupValidations
class, but it didn't fire on changes to the internal error tracking
structure. Not a big loss here - the form still won't submit if validation does not pass.