Add Nightly a11y Check
We now have access to setting a ENABLE_A11Y_AUDIT variable that we can
use to run an audit on a bunch of events. Let's enable this on a nightly
job and watch stuff fail, once we fix all the issues we can enable this
for all CI and remove the audits from all acceptance tests.