adds a11y attributes for indicating form validation errors.
i'm leaving out the learning material uploader out of this, since the
input element that would receive these a11y attrs is hidden. making this
work right would require another refactoring, but that would blow up the
scope of this change, so i'm punting on it for the time being.