Adjust the Header Layout
In order to keep the logo in it's box and apply a width to the search
input it is necessary to layout the columns in the grid. This has the
tendency to throw off the logo, so I've applied rules there to keep it
where we want it and at a size that looks nice.