Replace Logo with Sunburst on Phone Sized Screens
At very small sizes it is impossible to constrain the logo from over
running the search box. Replacing it with a square sunburst image allows
us to work with a fixed sized column width and locks down all the other
measurements. As the container gets bigger we have room for both,
however the logo doesn't like to stay in it's box, so I had to
specifically limit it's width.