Replace Grid Areas with Lines
Using the grid areas we cannot hid the search box area when the box
isn't present which forces the use of padding to make space instead of a
row-gap. However if we lay out the grid a bit more manually we can take
advantage of that very use property to ensure the search box doesn't
bump up against the other tools on small screens.