Add Support for Safari 17.x
In older safari versions a previous draft syntax of relative colors was
supports which required the unit to have a type attached. This syntax
isn't supported in other browsers or in Safari 18+. Using the
`@supports` syntax we can ensure we provide the right styles for each
browser until we drop Safari 17 support at the end of 2025.