scroll target element into center of the viewport.
by default, the scroll target will be top-aligned within the viewport.
in this case however, that top area is overlayed with a sticky table
header, which effectively would hide the scroll target element.
so i'm opting to verically center the scroll target instead, to make it
visible to the user.