removes misapplied aria labels from button.
according to SiteImprove, the accessible name of any interactive element should contain its visible text label.
which makes adding an aria-label attribute to these buttons pointless.

i checked into aria-description as an alternative, but according to MDN
the usage of that attribute is discouraged since it's still in Draft
mode in the spec.