Pin and Separate Days in Week at a Glance
It can be difficult in a busy week to see the split between days, and in
a busy day as you scroll to remember where you are on the page. Adding a
border to split the days makes it easier and as you scroll down the page
the day of the week pins itself to the top. This allows us to remove the
day from each even making them easier to read as well.