Improve a11y of Monthly Calendar
Give a little bit more room to the Show More button when it's present to
ensure it is clickable.

Add a minimum width to calendar boxes, which makes the screen break a
bit on extremely small windows, but ensures the days are still clickable
to users can navigate to them and they're not hidden.