Improve Readability of Single Events
The spacing and mix of bold and italic type made the critical
information here difficult to read. I've removed most of that and used
the icons to call out the attributes. I moved the location into the same
visual space as the time as I think that makes it easier to understand
what it is.

I also got rid of the triple curlies and used a getter to process the
session description, this does the same thing, but doesn't require us to
ignore a linting error.