Hide Unmatched Sessions in Search
When we only have a course match, but the session is small enough that
just the course data match scores it -  displaying it is confusing for
users as the badge is missing and the data isn't part of the session.
Instead we can check for matchedIn and know if there was any data in the
session that matched the search terms and filter out any sessions with
only course data matching.