Use Modifier for Save and Close
Instead of passing options to the component and adding a handler to the
container we can use a local modifier to handle the save  and close
keyboard interactions and pass that back to consumers of EditableField.
This makes the API easier to understand as not applying the modifier
will prevent any keyboard handing and allows us to remove an invalid
interactive issue.