Refactor FadeText
Using contextual components we can curry the values we need for
controlling text fading and use them where we apply this component. I
also found a way using css mask-image to fade out the text with the
component itself instead of needing to do that everywhere we use it.

These things combined let us wrap the EditableField inside the fade text
so we get the editable behavior, but maintain the expand collapse
buttons outside of the editable button.

Taken all together we can remove a lot of duplicate code and rely solely
on the FadeText component to do this work without needing to duplicate
the HTML and styles elsewhere.