Add a Test Waiter to FadeText
The resizeObserver fires asynchronously in browser code so tests may run
before it has a chance to detect the size and send it. This can cause
intermittent test failures. I've added a test waiter to FadeText to
ensure that we've gotten at least one call to register the size of the
element before we test anything.