Replace Copy Loading with Session Loading
The session copy isn't the issue, anything within the session route that
triggers loading is going to get a weird jump because the course loader
shows up. This session loader just puts a very large blue square on the
screen, which is better than a jump, but still not great.