Remove Awkward Loading from SessionCopy
Because this thing was wrapped in a loading check it would initially
render and empty div, this caused the page to jump up. Then, an instant
later, to fill out. Even if things were pre-loaded the course and year
loaders still had a moment of not being fulfilled and this would cause a
jump in the page.

Instead I've removed most of the loading behavior, and put a disabled
guard on the years drop down as that is the only place where we have
genuine async. The course route loads all the courses already so there
isn't any waiting there.

I remove all the pre-loading from the copy route, it wasn't doing
anything useful.