streamline event generation by replacing input pre-processing with input processing in constructor.
all the data wrangling happens directly in the constructor now.
this eliminates the need to a factory in the user/school events service.
furthermore, Event.slug is now be cached getter as it should be - this required some re-shuffling of utility methods.

this required a decent amount of rewiring in integration test coverage where we
have to now use actual Event objects as component input instead of mock
object that were hand-rolled or provided by mirage.