revert style aggregation and make ilios-table-colors its own class.
this undoes the consolidation of the ilios-table-structure and
ilios-table-colors mixins into one ilios-table class.
turns out during refactoring that we a ton of components that use the table structure is applied but
without the colors, so these must remain separate.
this applies the colors class to all tables that need it.