Move Update Day
We apply transitive updates automatically, but it appears that npm worms are
more likely to spread on the weekend. Moving this to Tuesday instead.

This pushed back our dependabot day to Wednesday (so we don't try and do
both). Which is closer to release data on Friday then I like, but still
should give us plenty of time.