Build Model Relationships for MSW
MSW Data sort of assumes all your models will be in the same file. I
don't love that, and there is a nice way to handle this. You define
getters on the schemas so they can reference each other. Unfortunately
we'd still have to import each schema and that gets circular and breaks.
Instead I'm dynamically building the getters using Object.defineProperty
which let's us declare the getter and wrap it in zod.

I also cleaned up all the unused schema exports and wrapped them in nice
arrays in models.js which makes it much easier to reference them
dynamically.